adding exam and training

+++ src/main/java/autoagencyuser/demo/repository/RoleRep.java
@@ -1,7 +1,11 @@
 package autoagencyuser.demo.repository;
 
 import autoagencyuser.demo.model.Role;
+import org.hibernate.tool.hbm2ddl.SchemaExportTask;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRep extends JpaRepository<Role, Long>{
+import java.util.Set;
+
+public interface RoleRep extends JpaRepository<Role, Integer>{
+
 }
